+++ client/OrzMC/Launcher(MacOS)/Views/Version+.swift
@@ -1,0 +1,8 @@
+//
+//  Version+.swift
+//  OrzMC
+//
+//  Created by joker on 2024/4/28.
+//
+
+import Foundation
